FPU: ksCurFPUOwner instead of ksActiveFPUState
Trivial change, except for x86 virtualisation and bootup.

Normally the seL4_TCBFlag_fpuDisabled TCB flag is used to decide whether to
enable or disable the FPU. However, with x86 virtualisation there is only
one TCB. Use the task flag for the host and always enable FPU for the guest.

When either the host or the guest's FPU is loaded, ksCurFPUOwner will point
to our TCB. saveFpuState and loadFpuState check fpu_active and do the right
thing depending on the current FPU user. To make this work it was necessary
to pass arch_tcb_t to saveFpuState and loadFpuState.

Signed-off-by: Indan Zupancic <indan@nul.nu>