Remove mcsPreemptionPoint and IRQ checking
There is no reason to check for pending interrupts after failed or
preempted system calls. All it does is save a bit of time avoiding
exiting and entering the kernel when preempting long-running kernel
operations. The slowdown for such operations after this change is a
factor of:

(kernel exit + entry time) / WCET

But making the code more complicated and slowing down all syscalls
to speed up a corner case is not a good trade off.

There is no reason for mcsPreemptionPoint to do basic timekeeping,
normal kernel operation is sufficient if we don't try to bypass
kernel exit/entry.

Signed-off-by: Indan Zupancic <indan@nul.nu>